+++ src/main/java/fr/kolala/advancedlocate/command/ConfiguratorCommand.java
@@ -12,7 +12,7 @@ import net.minecraft.text.Text;
 public class ConfiguratorCommand {
 
     public static void register (CommandDispatcher<ServerCommandSource> dispatcher) {
-        ConfigHelper.createConfigFileIfNotExisting();
+        ConfigHelper.createConfigFile();
 
         for (String field : ConfigHelper.listFields()) {
             dispatcher.register(CommandManager.literal("advancedlocate").requires(source -> source.hasPermissionLevel(2)).then(CommandManager.literal("config")
@@ -25,7 +25,7 @@ public class ConfiguratorCommand {
     }
 
     private static int getIntValue(ServerCommandSource source, String field) {
-        source.sendFeedback(() -> Text.translatable("command.advanced_locate.config.get", field, String.valueOf(ConfigHelper.getInt(field))), false);
+        source.sendFeedback(() -> Text.translatable("command.advancedlocate.config.get", field, String.valueOf(ConfigHelper.getInt(field))), false);
 
         return 0;
     }
@@ -34,14 +34,14 @@ public class ConfiguratorCommand {
         JsonObject json = ConfigHelper.read();
 
         if (json == null) {
-            source.sendFeedback(() -> Text.translatable("command.advanced_locate.config.fail", field, value), false);
+            source.sendFeedback(() -> Text.translatable("command.advancedlocate.config.fail", field, value), false);
             AdvancedLocate.LOGGER.error("Couldn't get json config.");
             return 1;
         }
 
         json.addProperty(field, value);
         if (ConfigHelper.write(json)) {
-            source.sendFeedback(() -> Text.translatable("command.advanced_locate.config.success", field, value), false);
+            source.sendFeedback(() -> Text.translatable("command.advancedlocate.config.success", field, value), false);
             AdvancedLocate.LOGGER.info("Successfully changed config file.");
 
             // Reload structure commands
@@ -50,7 +50,7 @@ public class ConfiguratorCommand {
             return 0;
         }
         else {
-            source.sendFeedback(() -> Text.translatable("command.advanced_locate.config.fail", field, value), false);
+            source.sendFeedback(() -> Text.translatable("command.advancedlocate.config.fail", field, value), false);
             AdvancedLocate.LOGGER.info("Couldn't change config.");
 
             return 1;